backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
psycopg2-binary
bcrypt
python-dotenv
pytest
pydantic>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.25.0
pytest-asyncio>=0.20.0
sqlalchemy[asyncio]>=2.0.0